Refactor EditExpense to use CancelButtonExpenseTemplate for navigation and remove unused imports
